Use common API client utils from graphql-client library
